# backend/internal/handler/oapi/server.gen.go
// Package oapi provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.7.0 DO NOT EDIT.
package oapi

import (
	"fmt"
	"net/http"
	"time"

	"github.com/labstack/echo/v4"
	"github.com/oapi-codegen/runtime"
	openapi_types "github.com/oapi-codegen/runtime/types"
)

const (
	CookieAuthScopes cookieAuthContextKey = "cookieAuth.Scopes"
)

// Defines values for PlanDetailResponseStatus.
const (
	PlanDetailResponseStatusActive    PlanDetailResponseStatus = "active"
	PlanDetailResponseStatusCompleted PlanDetailResponseStatus = "completed"
	PlanDetailResponseStatusPaused    PlanDetailResponseStatus = "paused"
)

// Valid indicates whether the value is a known member of the PlanDetailResponseStatus enum.
func (e PlanDetailResponseStatus) Valid() bool {
	switch e {
	case PlanDetailResponseStatusActive:
		return true
	case PlanDetailResponseStatusCompleted:
		return true
	case PlanDetailResponseStatusPaused:
		return true
	default:
		return false
	}
}

// Defines values for PlanResponseStatus.
const (
	PlanResponseStatusActive    PlanResponseStatus = "active"
	PlanResponseStatusCompleted PlanResponseStatus = "completed"
	PlanResponseStatusPaused    PlanResponseStatus = "paused"
)

// Valid indicates whether the value is a known member of the PlanResponseStatus enum.
func (e PlanResponseStatus) Valid() bool {
	switch e {
	case PlanResponseStatusActive:
		return true
	case PlanResponseStatusCompleted:
		return true
	case PlanResponseStatusPaused:
		return true
	default:
		return false
	}
}

// Defines values for SkillResponseSource.
const (
	Ai     SkillResponseSource = "ai"
	Manual SkillResponseSource = "manual"
)

// Valid indicates whether the value is a known member of the SkillResponseSource enum.
func (e SkillResponseSource) Valid() bool {
	switch e {
	case Ai:
		return true
	case Manual:
		return true
	default:
		return false
	}
}

// AvailabilityItem defines model for AvailabilityItem.
type AvailabilityItem struct {
	// DayOfWeek Go time.Weekday 準拠（0=日曜, 1=月曜, ..., 6=土曜）
	DayOfWeek int `json:"day_of_week"`

	// Hours 勉強可能時間（h）。0.5刻み
	Hours float32 `json:"hours"`
}

// AvailabilityResponse defines model for AvailabilityResponse.
type AvailabilityResponse struct {
	Availability []AvailabilityItem `json:"availability"`
	WeeklyTotal  float32            `json:"weekly_total"`
}

// AvatarPresetResponse defines model for AvatarPresetResponse.
type AvatarPresetResponse struct {
	AvatarPresetId string `json:"avatar_preset_id"`
}

// AvatarResponse defines model for AvatarResponse.
type AvatarResponse struct {
	AvatarPresetId  string               `json:"avatar_preset_id"`
	SkillCategories []SkillCategoryCount `json:"skill_categories"`
	SkillCount      int                  `json:"skill_count"`
}

// CreatePlanRequest defines model for CreatePlanRequest.
type CreatePlanRequest struct {
	// TargetDate 目標期限（YYYY-MM-DD）。今日より後の日付
	TargetDate openapi_types.Date `json:"target_date"`
	Title      string             `json:"title"`
	TotalPages int                `json:"total_pages"`
}

// CreateSkillRequest defines model for CreateSkillRequest.
type CreateSkillRequest struct {
	Category string `json:"category"`
	Name     string `json:"name"`

	// TaskId タスクに紐付ける場合のみ指定
	TaskId *openapi_types.UUID `json:"task_id,omitempty"`
}

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	Error string `json:"error"`
}

// ExtractSkillsResponse defines model for ExtractSkillsResponse.
type ExtractSkillsResponse struct {
	SuggestedSkills []SuggestedSkill `json:"suggested_skills"`
}

// LoginRequest defines model for LoginRequest.
type LoginRequest struct {
	Email    openapi_types.Email `json:"email"`
	Password string              `json:"password"`
}

// MessageResponse defines model for MessageResponse.
type MessageResponse struct {
	Message string `json:"message"`
}

// PlanDetailResponse defines model for PlanDetailResponse.
type PlanDetailResponse struct {
	AiReview         *string                  `json:"ai_review,omitempty"`
	CreatedAt        time.Time                `json:"created_at"`
	DailyPagesNeeded float32                  `json:"daily_pages_needed"`
	DaysRemaining    int                      `json:"days_remaining"`
	Id               openapi_types.UUID       `json:"id"`
	ProgressRate     float32                  `json:"progress_rate"`
	StartDate        openapi_types.Date       `json:"start_date"`
	Status           PlanDetailResponseStatus `json:"status"`
	TargetDate       openapi_types.Date       `json:"target_date"`
	Title            string                   `json:"title"`
	TotalPages       int                      `json:"total_pages"`
}

// PlanDetailResponseStatus defines model for PlanDetailResponse.Status.
type PlanDetailResponseStatus string

// PlanListResponse defines model for PlanListResponse.
type PlanListResponse struct {
	Plans []PlanResponse `json:"plans"`
}

// PlanResponse defines model for PlanResponse.
type PlanResponse struct {
	AiReview     *string            `json:"ai_review,omitempty"`
	CreatedAt    time.Time          `json:"created_at"`
	Id           openapi_types.UUID `json:"id"`
	ProgressRate float32            `json:"progress_rate"`
	StartDate    openapi_types.Date `json:"start_date"`
	Status       PlanResponseStatus `json:"status"`
	TargetDate   openapi_types.Date `json:"target_date"`
	Title        string             `json:"title"`
	TotalPages   int                `json:"total_pages"`
}

// PlanResponseStatus defines model for PlanResponse.Status.
type PlanResponseStatus string

// PlanReviewResponse defines model for PlanReviewResponse.
type PlanReviewResponse struct {
	AvailableDays int     `json:"available_days"`
	DailyPages    float32 `json:"daily_pages"`
	ReviewMessage string  `json:"review_message"`
	TotalDays     int     `json:"total_days"`
}

// SignupRequest defines model for SignupRequest.
type SignupRequest struct {
	DisplayName string              `json:"display_name"`
	Email       openapi_types.Email `json:"email"`
	Password    string              `json:"password"`
}

// SkillCategoryCount defines model for SkillCategoryCount.
type SkillCategoryCount struct {
	Category string `json:"category"`
	Count    int    `json:"count"`
}

// SkillListResponse defines model for SkillListResponse.
type SkillListResponse struct {
	Skills     []SkillResponse `json:"skills"`
	TotalCount int             `json:"total_count"`
}

// SkillResponse defines model for SkillResponse.
type SkillResponse struct {
	Category  string              `json:"category"`
	CreatedAt time.Time           `json:"created_at"`
	Id        openapi_types.UUID  `json:"id"`
	Name      string              `json:"name"`
	Source    SkillResponseSource `json:"source"`
	TaskId    *openapi_types.UUID `json:"task_id,omitempty"`
}

// SkillResponseSource defines model for SkillResponse.Source.
type SkillResponseSource string

// SuggestedSkill defines model for SuggestedSkill.
type SuggestedSkill struct {
	Category string `json:"category"`
	Name     string `json:"name"`
}

// UpdateAvailabilityRequest defines model for UpdateAvailabilityRequest.
type UpdateAvailabilityRequest struct {
	Availability []AvailabilityItem `json:"availability"`
}

// UpdateAvatarRequest defines model for UpdateAvatarRequest.
type UpdateAvatarRequest struct {
	// AvatarPresetId preset_01 〜 preset_12
	AvatarPresetId string `json:"avatar_preset_id"`
}

// UpdateSkillRequest defines model for UpdateSkillRequest.
type UpdateSkillRequest struct {
	Category string `json:"category"`
	Name     string `json:"name"`
}

// UserResponse defines model for UserResponse.
type UserResponse struct {
	DisplayName string             `json:"display_name"`
	Email       string             `json:"email"`
	Id          openapi_types.UUID `json:"id"`
}

// cookieAuthContextKey is the context key for cookieAuth security scheme
type cookieAuthContextKey string

// ListSkillsParams defines parameters for ListSkills.
type ListSkillsParams struct {
	// Category カテゴリでフィルタ
	Category *string `form:"category,omitempty" json:"category,omitempty"`
}

// LoginJSONRequestBody defines body for Login for application/json ContentType.
type LoginJSONRequestBody = LoginRequest

// SignupJSONRequestBody defines body for Signup for application/json ContentType.
type SignupJSONRequestBody = SignupRequest

// UpdateAvailabilityJSONRequestBody defines body for UpdateAvailability for application/json ContentType.
type UpdateAvailabilityJSONRequestBody = UpdateAvailabilityRequest

// UpdateAvatarJSONRequestBody defines body for UpdateAvatar for application/json ContentType.
type UpdateAvatarJSONRequestBody = UpdateAvatarRequest

// CreatePlanJSONRequestBody defines body for CreatePlan for application/json ContentType.
type CreatePlanJSONRequestBody = CreatePlanRequest

// ReviewPlanJSONRequestBody defines body for ReviewPlan for application/json ContentType.
type ReviewPlanJSONRequestBody = CreatePlanRequest

// CreateSkillJSONRequestBody defines body for CreateSkill for application/json ContentType.
type CreateSkillJSONRequestBody = CreateSkillRequest

// UpdateSkillJSONRequestBody defines body for UpdateSkill for application/json ContentType.
type UpdateSkillJSONRequestBody = UpdateSkillRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// ログイン
	// (POST /auth/login)
	Login(ctx echo.Context) error
	// ログアウト
	// (POST /auth/logout)
	Logout(ctx echo.Context) error
	// アクセストークンを更新
	// (POST /auth/refresh)
	RefreshToken(ctx echo.Context) error
	// ユーザー新規登録
	// (POST /auth/signup)
	Signup(ctx echo.Context) error
	// 可処分時間設定を取得
	// (GET /availability)
	GetAvailability(ctx echo.Context) error
	// 可処分時間設定を一括更新
	// (PUT /availability)
	UpdateAvailability(ctx echo.Context) error
	// アバター設定とスキルサマリーを取得
	// (GET /avatar)
	GetAvatar(ctx echo.Context) error
	// アバターのプリセットを設定/変更
	// (PUT /avatar)
	UpdateAvatar(ctx echo.Context) error
	// ヘルスチェック
	// (GET /health)
	HealthCheck(ctx echo.Context) error
	// 計画一覧を取得
	// (GET /plans)
	ListPlans(ctx echo.Context) error
	// 計画を作成
	// (POST /plans)
	CreatePlan(ctx echo.Context) error
	// 計画をAIにレビューしてもらう（保存しない）
	// (POST /plans/review)
	ReviewPlan(ctx echo.Context) error
	// 計画を削除
	// (DELETE /plans/{plan_id})
	DeletePlan(ctx echo.Context, planId openapi_types.UUID) error
	// 計画の詳細を取得
	// (GET /plans/{plan_id})
	GetPlan(ctx echo.Context, planId openapi_types.UUID) error
	// ユーザーの全スキル一覧を取得
	// (GET /skills)
	ListSkills(ctx echo.Context, params ListSkillsParams) error
	// スキルを手動追加
	// (POST /skills)
	CreateSkill(ctx echo.Context) error
	// スキルを削除
	// (DELETE /skills/{skill_id})
	DeleteSkill(ctx echo.Context, skillId openapi_types.UUID) error
	// スキルの名前やカテゴリを更新
	// (PUT /skills/{skill_id})
	UpdateSkill(ctx echo.Context, skillId openapi_types.UUID) error
	// タスクのメモからAIがスキルを抽出（保存しない）
	// (POST /tasks/{task_id}/extract-skills)
	ExtractSkills(ctx echo.Context, taskId openapi_types.UUID) error
}

// ServerInterfaceWrapper converts echo contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler ServerInterface
}

// Login converts echo context to params.
func (w *ServerInterfaceWrapper) Login(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.Login(ctx)
	return err
}

// Logout converts echo context to params.
func (w *ServerInterfaceWrapper) Logout(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.Logout(ctx)
	return err
}

// RefreshToken converts echo context to params.
func (w *ServerInterfaceWrapper) RefreshToken(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.RefreshToken(ctx)
	return err
}

// Signup converts echo context to params.
func (w *ServerInterfaceWrapper) Signup(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.Signup(ctx)
	return err
}

// GetAvailability converts echo context to params.
func (w *ServerInterfaceWrapper) GetAvailability(ctx echo.Context) error {
	var err error

	ctx.Set(string(CookieAuthScopes), []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetAvailability(ctx)
	return err
}

// UpdateAvailability converts echo context to params.
func (w *ServerInterfaceWrapper) UpdateAvailability(ctx echo.Context) error {
	var err error

	ctx.Set(string(CookieAuthScopes), []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.UpdateAvailability(ctx)
	return err
}

// GetAvatar converts echo context to params.
func (w *ServerInterfaceWrapper) GetAvatar(ctx echo.Context) error {
	var err error

	ctx.Set(string(CookieAuthScopes), []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetAvatar(ctx)
	return err
}

// UpdateAvatar converts echo context to params.
func (w *ServerInterfaceWrapper) UpdateAvatar(ctx echo.Context) error {
	var err error

	ctx.Set(string(CookieAuthScopes), []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.UpdateAvatar(ctx)
	return err
}

// HealthCheck converts echo context to params.
func (w *ServerInterfaceWrapper) HealthCheck(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.HealthCheck(ctx)
	return err
}

// ListPlans converts echo context to params.
func (w *ServerInterfaceWrapper) ListPlans(ctx echo.Context) error {
	var err error

	ctx.Set(string(CookieAuthScopes), []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.ListPlans(ctx)
	return err
}

// CreatePlan converts echo context to params.
func (w *ServerInterfaceWrapper) CreatePlan(ctx echo.Context) error {
	var err error

	ctx.Set(string(CookieAuthScopes), []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.CreatePlan(ctx)
	return err
}

// ReviewPlan converts echo context to params.
func (w *ServerInterfaceWrapper) ReviewPlan(ctx echo.Context) error {
	var err error

	ctx.Set(string(CookieAuthScopes), []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.ReviewPlan(ctx)
	return err
}

// DeletePlan converts echo context to params.
func (w *ServerInterfaceWrapper) DeletePlan(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "plan_id" -------------
	var planId openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "plan_id", ctx.Param("plan_id"), &planId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "uuid"})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter plan_id: %s", err))
	}

	ctx.Set(string(CookieAuthScopes), []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.DeletePlan(ctx, planId)
	return err
}

// GetPlan converts echo context to params.
func (w *ServerInterfaceWrapper) GetPlan(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "plan_id" -------------
	var planId openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "plan_id", ctx.Param("plan_id"), &planId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "uuid"})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter plan_id: %s", err))
	}

	ctx.Set(string(CookieAuthScopes), []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetPlan(ctx, planId)
	return err
}

// ListSkills converts echo context to params.
func (w *ServerInterfaceWrapper) ListSkills(ctx echo.Context) error {
	var err error

	ctx.Set(string(CookieAuthScopes), []string{})

	// Parameter object where we will unmarshal all parameters from the context
	var params ListSkillsParams
	// ------------- Optional query parameter "category" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "category", ctx.QueryParams(), &params.Category, runtime.BindQueryParameterOptions{Type: "string", Format: ""})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter category: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.ListSkills(ctx, params)
	return err
}

// CreateSkill converts echo context to params.
func (w *ServerInterfaceWrapper) CreateSkill(ctx echo.Context) error {
	var err error

	ctx.Set(string(CookieAuthScopes), []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.CreateSkill(ctx)
	return err
}

// DeleteSkill converts echo context to params.
func (w *ServerInterfaceWrapper) DeleteSkill(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "skill_id" -------------
	var skillId openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "skill_id", ctx.Param("skill_id"), &skillId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "uuid"})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter skill_id: %s", err))
	}

	ctx.Set(string(CookieAuthScopes), []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.DeleteSkill(ctx, skillId)
	return err
}

// UpdateSkill converts echo context to params.
func (w *ServerInterfaceWrapper) UpdateSkill(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "skill_id" -------------
	var skillId openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "skill_id", ctx.Param("skill_id"), &skillId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "uuid"})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter skill_id: %s", err))
	}

	ctx.Set(string(CookieAuthScopes), []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.UpdateSkill(ctx, skillId)
	return err
}

// ExtractSkills converts echo context to params.
func (w *ServerInterfaceWrapper) ExtractSkills(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "task_id" -------------
	var taskId openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "task_id", ctx.Param("task_id"), &taskId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "uuid"})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter task_id: %s", err))
	}

	ctx.Set(string(CookieAuthScopes), []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.ExtractSkills(ctx, taskId)
	return err
}

// This is a simple interface which specifies echo.Route addition functions which
// are present on both echo.Echo and echo.Group, since we want to allow using
// either of them for path registration
type EchoRouter interface {
	CONNECT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	DELETE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	GET(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	HEAD(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	OPTIONS(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PATCH(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	POST(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PUT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	TRACE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
}

// RegisterHandlersOptions configures RegisterHandlersWithOptions.
type RegisterHandlersOptions struct {
	// BaseURL is prepended to every registered path so the API can be served
	// under a prefix.
	BaseURL string
	// OperationMiddlewares lets the caller attach per-operation middleware at
	// registration time. The map key is the OpenAPI `operationId` value as it
	// appears in the spec (the raw, un-normalized form). Operations that have
	// no entry are registered with no extra middleware. A nil map disables
	// per-operation middleware entirely.
	OperationMiddlewares map[string][]echo.MiddlewareFunc
}

// RegisterHandlers adds each server route to the EchoRouter.
func RegisterHandlers(router EchoRouter, si ServerInterface) {
	RegisterHandlersWithOptions(router, si, RegisterHandlersOptions{})
}

// RegisterHandlersWithBaseURL registers handlers and prepends BaseURL to the
// paths so the API can be served under a prefix.
func RegisterHandlersWithBaseURL(router EchoRouter, si ServerInterface, baseURL string) {
	RegisterHandlersWithOptions(router, si, RegisterHandlersOptions{BaseURL: baseURL})
}

// RegisterHandlersWithOptions registers handlers using the supplied options,
// including any per-operation middleware.
func RegisterHandlersWithOptions(router EchoRouter, si ServerInterface, options RegisterHandlersOptions) {

	wrapper := ServerInterfaceWrapper{
		Handler: si,
	}

	router.POST(options.BaseURL+"/auth/login", wrapper.Login, options.OperationMiddlewares["login"]...)
	router.POST(options.BaseURL+"/auth/logout", wrapper.Logout, options.OperationMiddlewares["logout"]...)
	router.POST(options.BaseURL+"/auth/refresh", wrapper.RefreshToken, options.OperationMiddlewares["refreshToken"]...)
	router.POST(options.BaseURL+"/auth/signup", wrapper.Signup, options.OperationMiddlewares["signup"]...)
	router.GET(options.BaseURL+"/availability", wrapper.GetAvailability, options.OperationMiddlewares["getAvailability"]...)
	router.PUT(options.BaseURL+"/availability", wrapper.UpdateAvailability, options.OperationMiddlewares["updateAvailability"]...)
	router.GET(options.BaseURL+"/avatar", wrapper.GetAvatar, options.OperationMiddlewares["getAvatar"]...)
	router.PUT(options.BaseURL+"/avatar", wrapper.UpdateAvatar, options.OperationMiddlewares["updateAvatar"]...)
	router.GET(options.BaseURL+"/health", wrapper.HealthCheck, options.OperationMiddlewares["healthCheck"]...)
	router.GET(options.BaseURL+"/plans", wrapper.ListPlans, options.OperationMiddlewares["listPlans"]...)
	router.POST(options.BaseURL+"/plans", wrapper.CreatePlan, options.OperationMiddlewares["createPlan"]...)
	router.POST(options.BaseURL+"/plans/review", wrapper.ReviewPlan, options.OperationMiddlewares["reviewPlan"]...)
	router.DELETE(options.BaseURL+"/plans/:plan_id", wrapper.DeletePlan, options.OperationMiddlewares["deletePlan"]...)
	router.GET(options.BaseURL+"/plans/:plan_id", wrapper.GetPlan, options.OperationMiddlewares["getPlan"]...)
	router.GET(options.BaseURL+"/skills", wrapper.ListSkills, options.OperationMiddlewares["listSkills"]...)
	router.POST(options.BaseURL+"/skills", wrapper.CreateSkill, options.OperationMiddlewares["createSkill"]...)
	router.DELETE(options.BaseURL+"/skills/:skill_id", wrapper.DeleteSkill, options.OperationMiddlewares["deleteSkill"]...)
	router.PUT(options.BaseURL+"/skills/:skill_id", wrapper.UpdateSkill, options.OperationMiddlewares["updateSkill"]...)
	router.POST(options.BaseURL+"/tasks/:task_id/extract-skills", wrapper.ExtractSkills, options.OperationMiddlewares["extractSkills"]...)

}