# docs/openapi/openapi.yaml
openapi: "3.0.3"
info:
  title: Ninjo API
  version: "1.0.0"
  description: 勉強計画を立て、AIと一緒に「やり切る」を実現するAPI
servers:
  - url: /api/v1
paths:
  /auth/signup:
    post:
      operationId: signup
      summary: ユーザー新規登録
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupRequest"
      responses:
        "201":
          description: 登録成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: メールアドレスが既に登録されている
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      summary: ログイン
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: ログイン成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          description: メールアドレスまたはパスワードが間違っている
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/refresh:
    post:
      operationId: refreshToken
      summary: アクセストークンを更新
      tags: [auth]
      responses:
        "200":
          description: トークン更新成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          description: リフレッシュトークンが無効または期限切れ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /avatar:
    get:
      operationId: getAvatar
      summary: アバター設定とスキルサマリーを取得
      tags: [avatar]
      security:
        - cookieAuth: []
      responses:
        "200":
          description: 取得成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvatarResponse"
        "401":
          description: 未認証
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateAvatar
      summary: アバターのプリセットを設定/変更
      tags: [avatar]
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAvatarRequest"
      responses:
        "200":
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvatarPresetResponse"
        "400":
          description: 無効なプリセットID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: 未認証
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /availability:
    get:
      operationId: getAvailability
      summary: 可処分時間設定を取得
      tags: [availability]
      security:
        - cookieAuth: []
      responses:
        "200":
          description: 取得成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvailabilityResponse"
        "401":
          description: 未認証
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateAvailability
      summary: 可処分時間設定を一括更新
      tags: [availability]
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAvailabilityRequest"
      responses:
        "200":
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvailabilityResponse"
        "400":
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: 未認証
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /plans:
    get:
      operationId: listPlans
      summary: 計画一覧を取得
      tags: [plan]
      security:
        - cookieAuth: []
      responses:
        "200":
          description: 取得成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanListResponse"
        "401":
          description: 未認証
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createPlan
      summary: 計画を作成
      tags: [plan]
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePlanRequest"
      responses:
        "201":
          description: 作成成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanResponse"
        "400":
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: 未認証
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /plans/review:
    post:
      operationId: reviewPlan
      summary: 計画をAIにレビューしてもらう（保存しない）
      tags: [plan]
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePlanRequest"
      responses:
        "200":
          description: レビュー結果
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanReviewResponse"
        "400":
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: 未認証
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /plans/{plan_id}:
    get:
      operationId: getPlan
      summary: 計画の詳細を取得
      tags: [plan]
      security:
        - cookieAuth: []
      parameters:
        - name: plan_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: 取得成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanDetailResponse"
        "401":
          description: 未認証
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: 計画が見つからない
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deletePlan
      summary: 計画を削除
      tags: [plan]
      security:
        - cookieAuth: []
      parameters:
        - name: plan_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: 削除成功
        "401":
          description: 未認証
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: 計画が見つからない
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /tasks/{task_id}/extract-skills:
    post:
      operationId: extractSkills
      summary: タスクのメモからAIがスキルを抽出（保存しない）
      tags: [skill]
      security:
        - cookieAuth: []
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: 抽出成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExtractSkillsResponse"
        "400":
          description: メモが空
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: 未認証
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: タスクが存在しない
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
    SignupRequest:
      type: object
      required: [email, password, display_name]
      properties:
        email:
          type: string
          format: email
          maxLength: 255
        password:
          type: string
          minLength: 8
        display_name:
          type: string
          minLength: 1
          maxLength: 100
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    UserResponse:
      type: object
      required: [id, email, display_name]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        display_name:
          type: string
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
...
    AvailabilityResponse:
      type: object
      required: [availability, weekly_total]
      properties:
        availability:
          type: array
          items:
            $ref: "#/components/schemas/AvailabilityItem"
          minItems: 7
          maxItems: 7
        weekly_total:
          type: number
          format: float
    UpdateAvailabilityRequest:
      type: object
      required: [availability]
      properties:
        availability:
          type: array
          items:
            $ref: "#/components/schemas/AvailabilityItem"
          minItems: 7
          maxItems: 7
    AvatarResponse:
      type: object
      required: [avatar_preset_id, skill_count, skill_categories]
      properties:
        avatar_preset_id:
          type: string
        skill_count:
          type: integer
        skill_categories:
          type: array
          items:
            $ref: "#/components/schemas/SkillCategoryCount"
...
    UpdateAvatarRequest:
      type: object
      required: [avatar_preset_id]
      properties:
        avatar_preset_id:
          type: string
          description: "preset_01 〜 preset_12"
    AvatarPresetResponse:
      type: object
      required: [avatar_preset_id]
      properties:
        avatar_preset_id:
          type: string
...
    PlanListResponse:
      type: object
      required: [plans]
      properties:
        plans:
          type: array
          items:
            $ref: "#/components/schemas/PlanResponse"
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token